requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
# Pour production:
# Pillow==11.0.0
# psycopg2-binary==2.9.9
